Sounds and better collisions
Collisions can get sticky in Physics2D. This solves that problem, there
are also some sounds that are playing now that kind of suck but it's
something to start from. Also there is the ability for the computer to
play though it would be much nicer if it could play through all the
levels without intervention. Basically, I'm thinking of attaching a
different game object that would handle it if necessary.