LIVES WORK!
Got lives working. Not the most elegant wiring up, but it is getting
closer to the intention I believe of the ECS that unity provides. I can
see that if you have a good understanding of what states you are trying
to accomplish you can really use minimal code to accomplish objectives.